Added extractZipFile functionality
